requirements.txt
# Bentley Budget Bot - Streamlit Cloud Requirements
# Optimized for Streamlit Cloud deployment at bbbot305.streamlit.app
# For local development with full features, see requirements-local.txt
# Updated: December 25, 2025 - Fixed MySQL connector and environment loading

# ============================================
# CORE STREAMLIT APPLICATION
# ============================================

# Streamlit framework
streamlit>=1.28.0

# Data visualization
plotly>=5.0.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.3.0

# Financial data
yfinance>=0.2.0,<0.2.66
websockets>=9,<11

# Kalshi prediction markets - Official SDK
kalshi>=0.2.0

# Web scraping
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP requests
requests>=2.28.0

# MT5 REST bridge - FastAPI (supports both local and Railway deployment)
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

# Legacy Flask support (optional)
flask>=3.0.0
flask-cors>=5.0.0
gunicorn>=23.0.0

# Appwrite backend SDK
appwrite>=4.0.0

# ============================================
# CACHING & AI (Google Generative AI)
# ============================================
# Pin cachetools to compatible version
# Prevents ImportError: cachetools.func on cloud
cachetools>=5.0.0,<6.0.0
google-generativeai>=0.3.0

# ============================================
# PLAID INTEGRATION (Personal Budget Analysis)
# ============================================
plaid-python==30.0.0
cryptography>=46.0.0
python-dateutil>=2.8.0

# ============================================
# PREDICTION MARKETS (Kalshi & Polymarket)
# ============================================
cryptography>=46.0.0  # RSA signing for Kalshi API authentication
aiohttp>=3.9.0

# ============================================
# DATABASE CONNECTIVITY (Required for Budget Features)
# ============================================
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0
pymysql>=1.1.0

# ============================================
# MLFLOW EXPERIMENT TRACKING (Required for Investment Analysis)
# ============================================
mlflow>=2.15.0,<3.0.0
alembic>=1.12.0

# ============================================
# AI CHATBOT (Bentley AI Assistant)
# ============================================
# Note: DeepSeek or other LLM APIs require API keys
# Add to .env: DEEPSEEK_API_KEY=your_key_here

# ============================================
# NOTE: Heavy packages excluded for Streamlit Cloud
# ============================================
# For local development with full pipeline features, use:
# pip install -r requirements-local.txt
#
# Excluded from cloud deployment:
# - apache-airflow (too large, conflicts)
# - dbt-core/dbt-mysql (local transformation only)
# - broker trading SDKs (local trading only - APIs work via REST)
# 
# INCLUDED for cloud (lightweight):
# - mlflow (experiment tracking for Investment Analysis page)
# - appwrite (backend SDK for database operations)
# ============================================
